Homework 8. Small refactoring
