+++ Game/index.js
@@ -1,8 +1,8 @@
-import { getRandom, createElement, getTime } from "../utils/index.js";
-import { HIT, ATTACK, LOGS } from "../constants/index.js";
-import { Scorpion, Subzero, createPlayer } from "../Player/index.js";
+import { getRandom, createElement, getTime } from "../utils";
+import { HIT, ATTACK, LOGS } from "../constants";
+import { Scorpion, Subzero, createPlayer } from "../Player";
 
-export default class Game {
+export class Game {
     constructor() {
         this.formFight = document.querySelector('.control');
         this.arenas = document.querySelector('.arenas');
@@ -87,28 +87,22 @@ export default class Game {
                     .replace('[player1]', playerName1)
                     .replace('[player2]', playerName2)
                     .replace('[time]', getTime());
-                break;
             case 'hit':
                 return LOGS[type][getRandom(LOGS[type].length - 1) - 1]
                     .replace('[playerKick]', playerName1)
                     .replace('[playerDefence]', playerName2);
-                break;
             case 'defence':
                 return LOGS[type][getRandom(LOGS[type].length - 1) - 1]
                     .replace('[playerKick]', playerName1)
                     .replace('[playerDefence]', playerName2);
-                break;
             case 'end':
                 return LOGS[type][getRandom(LOGS[type].length - 1) - 1]
                     .replace('[playerWins]', playerName1)
                     .replace('[playerLose]', playerName2);
-                break;
             case 'draw':
                 return LOGS[type];
-                break;
             default:
                 return 'Something else';
-                break;
         }      
     }
