+++ main.js
@@ -1,4 +1,4 @@
-import Game from "../Game/index.js";
+import { Game } from "./Game";
 
 const game = new Game();
